# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: OKSTours AI Agent API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: conversations
    description: Chat conversation management
  - name: messages
    description: Chat messages
  - name: leads
    description: Lead/customer management
  - name: chat
    description: AI chat operations
  - name: stats
    description: Statistics and analytics
  - name: promotions
    description: Promo and discount management
  - name: callanalysis
    description: Call recording analysis
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /conversations:
    get:
      operationId: listConversations
      tags: [conversations]
      summary: List all conversations
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [active, closed, pending]
        - name: channel
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Conversation"
    post:
      operationId: createConversation
      tags: [conversations]
      summary: Create a new conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateConversationBody"
      responses:
        "201":
          description: Created conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"

  /conversations/{id}:
    get:
      operationId: getConversation
      tags: [conversations]
      summary: Get a conversation by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Conversation with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConversationWithMessages"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateConversation
      tags: [conversations]
      summary: Update a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateConversationBody"
      responses:
        "200":
          description: Updated conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /conversations/{id}/messages:
    get:
      operationId: getConversationMessages
      tags: [messages]
      summary: Get messages for a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"

  /conversations/{id}/operator-reply:
    post:
      operationId: operatorReply
      tags: [conversations]
      summary: Operator sends a message in a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OperatorReplyBody"
      responses:
        "200":
          description: Operator message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /conversations/{id}/follow-up:
    post:
      operationId: sendFollowUp
      tags: [conversations]
      summary: Send an automatic follow-up message
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Follow-up sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /chat:
    post:
      operationId: sendMessage
      tags: [chat]
      summary: Send a message and get AI response
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageBody"
      responses:
        "200":
          description: AI response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /leads:
    get:
      operationId: listLeads
      tags: [leads]
      summary: List all leads/customers
      parameters:
        - name: segment
          in: query
          required: false
          schema:
            type: string
            enum: [hot, warm, cold]
        - name: interest
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of leads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lead"
    post:
      operationId: createLead
      tags: [leads]
      summary: Create a new lead
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLeadBody"
      responses:
        "201":
          description: Created lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"

  /leads/{id}:
    get:
      operationId: getLead
      tags: [leads]
      summary: Get a lead by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateLead
      tags: [leads]
      summary: Update a lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLeadBody"
      responses:
        "200":
          description: Updated lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /leads/{id}/book:
    post:
      operationId: bookLead
      tags: [leads]
      summary: Confirm booking for a lead and send confirmation message
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookLeadBody"
      responses:
        "200":
          description: Booking confirmed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /promotions:
    get:
      operationId: listPromotions
      tags: [promotions]
      summary: List all active promotions
      responses:
        "200":
          description: List of promotions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Promotion"
    post:
      operationId: createPromotion
      tags: [promotions]
      summary: Create a new promotion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePromotionBody"
      responses:
        "201":
          description: Created promotion
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Promotion"

  /promotions/{id}:
    delete:
      operationId: deletePromotion
      tags: [promotions]
      summary: Delete a promotion
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /call-analysis:
    post:
      operationId: analyzeCall
      tags: [callanalysis]
      summary: Analyze a call transcript or audio text
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalyzeCallBody"
      responses:
        "200":
          description: Call analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CallAnalysisResult"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /stats/dashboard:
    get:
      operationId: getDashboardStats
      tags: [stats]
      summary: Get dashboard statistics
      responses:
        "200":
          description: Dashboard statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"

  /stats/activity:
    get:
      operationId: getRecentActivity
      tags: [stats]
      summary: Get recent activity feed
      responses:
        "200":
          description: Recent activity
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityItem"

  /stats/time-series:
    get:
      operationId: getTimeSeries
      tags: [stats]
      summary: Get time-series stats (daily/weekly/monthly)
      parameters:
        - name: period
          in: query
          required: false
          schema:
            type: string
            enum: [daily, weekly, monthly]
      responses:
        "200":
          description: Time series data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimeSeriesStats"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    Conversation:
      type: object
      properties:
        id:
          type: integer
        channel:
          type: string
          enum: [telegram, whatsapp, instagram, web, sms, email]
        status:
          type: string
          enum: [active, closed, pending]
        customerName:
          type: ["string", "null"]
        customerPhone:
          type: ["string", "null"]
        leadId:
          type: ["integer", "null"]
        lastMessage:
          type: ["string", "null"]
        lastMessageAt:
          type: ["string", "null"]
        operatorMode:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - channel
        - status
        - customerName
        - customerPhone
        - leadId
        - lastMessage
        - lastMessageAt
        - operatorMode
        - createdAt
        - updatedAt

    ConversationWithMessages:
      type: object
      properties:
        id:
          type: integer
        channel:
          type: string
          enum: [telegram, whatsapp, instagram, web, sms, email]
        status:
          type: string
          enum: [active, closed, pending]
        customerName:
          type: ["string", "null"]
        customerPhone:
          type: ["string", "null"]
        leadId:
          type: ["integer", "null"]
        lastMessage:
          type: ["string", "null"]
        lastMessageAt:
          type: ["string", "null"]
        operatorMode:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
        messages:
          type: array
          items:
            $ref: "#/components/schemas/Message"
      required:
        - id
        - channel
        - status
        - customerName
        - customerPhone
        - leadId
        - lastMessage
        - lastMessageAt
        - operatorMode
        - createdAt
        - updatedAt
        - messages

    CreateConversationBody:
      type: object
      properties:
        channel:
          type: string
          enum: [telegram, whatsapp, instagram, web, sms, email]
        customerName:
          type: ["string", "null"]
        customerPhone:
          type: ["string", "null"]
      required:
        - channel

    UpdateConversationBody:
      type: object
      properties:
        status:
          type: string
          enum: [active, closed, pending]
        customerName:
          type: ["string", "null"]
        customerPhone:
          type: ["string", "null"]
        leadId:
          type: ["integer", "null"]
        operatorMode:
          type: boolean

    OperatorReplyBody:
      type: object
      properties:
        content:
          type: string
        operatorName:
          type: ["string", "null"]
      required:
        - content

    Message:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        role:
          type: string
          enum: [user, assistant, operator]
        content:
          type: string
        createdAt:
          type: string
      required:
        - id
        - conversationId
        - role
        - content
        - createdAt

    SendMessageBody:
      type: object
      properties:
        conversationId:
          type: ["integer", "null"]
        message:
          type: string
        channel:
          type: string
          enum: [telegram, whatsapp, instagram, web, sms, email]
        customerName:
          type: ["string", "null"]
      required:
        - message
        - channel

    ChatResponse:
      type: object
      properties:
        conversationId:
          type: integer
        message:
          type: string
        leadCreated:
          type: boolean
      required:
        - conversationId
        - message
        - leadCreated

    Lead:
      type: object
      properties:
        id:
          type: integer
        name:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        segment:
          type: string
          enum: [hot, warm, cold]
        interest:
          type: ["string", "null"]
        destination:
          type: ["string", "null"]
        budget:
          type: ["string", "null"]
        status:
          type: string
          enum: [new, contacted, qualified, booked, lost]
        notes:
          type: ["string", "null"]
        conversationId:
          type: ["integer", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - phone
        - email
        - segment
        - interest
        - destination
        - budget
        - status
        - notes
        - conversationId
        - createdAt
        - updatedAt

    CreateLeadBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        segment:
          type: string
          enum: [hot, warm, cold]
        interest:
          type: ["string", "null"]
        destination:
          type: ["string", "null"]
        budget:
          type: ["string", "null"]
        status:
          type: string
          enum: [new, contacted, qualified, booked, lost]
        notes:
          type: ["string", "null"]
        conversationId:
          type: ["integer", "null"]
      required:
        - segment
        - status

    UpdateLeadBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        segment:
          type: string
          enum: [hot, warm, cold]
        interest:
          type: ["string", "null"]
        destination:
          type: ["string", "null"]
        budget:
          type: ["string", "null"]
        status:
          type: string
          enum: [new, contacted, qualified, booked, lost]
        notes:
          type: ["string", "null"]

    BookLeadBody:
      type: object
      properties:
        travelDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]

    Promotion:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        discount:
          type: ["string", "null"]
        destination:
          type: ["string", "null"]
        validUntil:
          type: ["string", "null"]
        active:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - title
        - description
        - discount
        - destination
        - validUntil
        - active
        - createdAt

    CreatePromotionBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        discount:
          type: ["string", "null"]
        destination:
          type: ["string", "null"]
        validUntil:
          type: ["string", "null"]
      required:
        - title
        - description

    AnalyzeCallBody:
      type: object
      properties:
        transcript:
          type: string
        audioBase64:
          type: ["string", "null"]
        language:
          type: string
          enum: [uz, ru, en]
      required:
        - transcript

    CallAnalysisResult:
      type: object
      properties:
        summary:
          type: string
        clientRequest:
          type: string
        objections:
          type: string
        missedOpportunities:
          type: string
        recommendations:
          type: string
        sentiment:
          type: string
          enum: [positive, neutral, negative]
        leadQuality:
          type: string
          enum: [hot, warm, cold]
      required:
        - summary
        - clientRequest
        - objections
        - missedOpportunities
        - recommendations
        - sentiment
        - leadQuality

    DashboardStats:
      type: object
      properties:
        totalConversations:
          type: integer
        activeConversations:
          type: integer
        totalLeads:
          type: integer
        hotLeads:
          type: integer
        warmLeads:
          type: integer
        coldLeads:
          type: integer
        bookedLeads:
          type: integer
        channelBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/ChannelCount"
        todayMessages:
          type: integer
      required:
        - totalConversations
        - activeConversations
        - totalLeads
        - hotLeads
        - warmLeads
        - coldLeads
        - bookedLeads
        - channelBreakdown
        - todayMessages

    ChannelCount:
      type: object
      properties:
        channel:
          type: string
        count:
          type: integer
      required:
        - channel
        - count

    ActivityItem:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [new_lead, new_message, booking, status_change, follow_up, operator_reply]
        description:
          type: string
        conversationId:
          type: ["integer", "null"]
        leadId:
          type: ["integer", "null"]
        createdAt:
          type: string
      required:
        - id
        - type
        - description
        - conversationId
        - leadId
        - createdAt

    TimeSeriesStats:
      type: object
      properties:
        period:
          type: string
        data:
          type: array
          items:
            $ref: "#/components/schemas/TimeSeriesPoint"
        totalMessages:
          type: integer
        totalLeads:
          type: integer
        totalBookings:
          type: integer
      required:
        - period
        - data
        - totalMessages
        - totalLeads
        - totalBookings

    TimeSeriesPoint:
      type: object
      properties:
        date:
          type: string
        messages:
          type: integer
        leads:
          type: integer
        bookings:
          type: integer
      required:
        - date
        - messages
        - leads
        - bookings