# lib/api-zod/src/generated/types/analyzeCallBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * OKSTours AI Agent API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AnalyzeCallBodyLanguage } from "./analyzeCallBodyLanguage";

export interface AnalyzeCallBody {
  transcript: string;
  /** @nullable */
  audioBase64?: string | null;
  language?: AnalyzeCallBodyLanguage;
}