# lib/api-zod/src/generated/types/createPromotionBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * OKSTours AI Agent API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreatePromotionBody {
  title: string;
  description: string;
  /** @nullable */
  discount?: string | null;
  /** @nullable */
  destination?: string | null;
  /** @nullable */
  validUntil?: string | null;
}