# lib/api-zod/src/generated/types/operatorReplyBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * OKSTours AI Agent API specification
 * OpenAPI spec version: 0.1.0
 */

export interface OperatorReplyBody {
  content: string;
  /** @nullable */
  operatorName?: string | null;
}