# lib/api-zod/src/generated/types/timeSeriesPoint.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * OKSTours AI Agent API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TimeSeriesPoint {
  date: string;
  messages: number;
  leads: number;
  bookings: number;
}