# lib/api-zod/src/generated/types/timeSeriesStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * OKSTours AI Agent API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TimeSeriesPoint } from "./timeSeriesPoint";

export interface TimeSeriesStats {
  period: string;
  data: TimeSeriesPoint[];
  totalMessages: number;
  totalLeads: number;
  totalBookings: number;
}